feat: Add optional parameter for registerDirty to specify exactly what fields need to be updated
This allows a unit of work's uncommited changes to not be overwritten when multiple areas of code update the same record on the same unit of work.

e.g. a stack of processes that update different fields on the same record while being passed a unit of work allowing for changes to be commited after http callouts

Comes form Traction on Demand's John Rogers, has been sitting in our code base for a long time

There is a breaking change compared to a previous versions of fflib, where an excpetion is thrown if the same record is registered dirty twice without fields being specified. This could be modified to overwrite the existing registered dirty record (which is the current behaviour) instead of throwing an exception.